% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_report.R
\name{paged_report}
\alias{paged_report}
\title{Function for RMS report paged template}
\usage{
paged_report(
  other_css = NULL,
  toc = TRUE,
  toc_depth = 2,
  number_sections = FALSE,
  ...
)
}
\arguments{
\item{other_css}{Add extra css}

\item{toc}{Table of content}

\item{toc_depth}{Table of content depth}

\item{number_sections}{Number section headings}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown report
}
\description{
Function for RMS report paged template
}
