% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_bar_chart.R
\name{simple_bar_chart}
\alias{simple_bar_chart}
\title{Create a simple horizontal bar chart}
\usage{
simple_bar_chart(
  df,
  cat,
  val,
  title = "Bar Chart",
  caption = "Notes",
  chr_wrap = NULL,
  percentage = FALSE,
  bar_color = NULL,
  lbl_threshold = NULL
)
}
\arguments{
\item{df}{Formatted data frame}

\item{cat}{Category column}

\item{val}{Value column}

\item{title}{Chart title}

\item{caption}{Chart caption}

\item{chr_wrap}{Number of character to wrap long category labels}

\item{percentage}{Format axis and labels as percentage TRUE/FALSE}

\item{bar_color}{Specific color code for bars}

\item{lbl_threshold}{Value to switch labels from inside to outside the bars}
}
\value{
A plot object
}
\description{
Create a simple horizontal bar chart
}
\examples{
\dontrun{

library(ggplot2)

df_percent <- data.frame(
  category = c("Category A", "Category B", "Category C", "Category D"),
  value = c(0.3, 0.08, 0.6, 0.75)
)

simple_bar_chart(df_percent, cat = "category", val = "value")
}
}
